Moved fitting modules to fitting directory
