+++ slac_measurements/fitting/gaussian.py
@@ -1,6 +1,6 @@
 import numpy as np
 from scipy.stats import norm
-import slac_measurements.image.optimize as optimize
+import slac_measurements.fitting.optimize as optimize
 
 
 def curve(x, mean=0, sigma=1, amp=1, off=0):
